+++ resources/views/dashboard/education.blade.php
@@ -28,6 +28,7 @@
                 <div class="card-header">{{ __('EDUCATION') }}</div>
 
                 <div class="card-body">
+                    <x-alert />
                     <table>
                         <tr>
                             <th>ID</th>
@@ -52,8 +53,8 @@
                                 <td></td>
                             @endif
                             <td class="even actions">
-                                <a href='education/edit/{{ $d->id }}'>Edit</a>
-                                <a href='education/delete/{{ $d->id }}'>Delete</a>
+                                <a class="btn btn-outline-primary" href='education/edit/{{ $d->id }}'>Edit</a>
+                                <a class="btn btn-outline-danger" onclick="return confirm('Are you sure?')" href='education/delete/{{ $d->id }}'>Delete</a>
                             </td>
                         </tr> 
                             
@@ -70,13 +71,13 @@
         <div class="card">
             <div class="card-header">{{ __('Create New Data') }}</div>
                 <div class="card-body">
-                    <form method="post" action="/dashboard/skill" enctype="multipart/form-data">
+                    <form method="post" action="/dashboard/education" enctype="multipart/form-data">
                         @csrf
                         <input type="text" name="school_name" placeholder="Name"/>
                         <input type="text" name="year_started" placeholder="Year Started"/>
                         <input type="text" name="year_graduated" placeholder="Year Graduated"/>
                         <input type="file" name="logo" />
-                        <button type="submit" class="submit">Create</button>
+                        <button type="submit" class="submit btn btn-outline-secondary">Create</button>
                     </form>
                 </div>
         </div>
